PR refactor
